Created transaction module. Moved files for screen from app to this module
