+++ .idea/gradle.xml
@@ -16,6 +16,7 @@
             <option value="$PROJECT_DIR$/core_impl" />
             <option value="$PROJECT_DIR$/general" />
             <option value="$PROJECT_DIR$/main" />
+            <option value="$PROJECT_DIR$/transactions" />
             <option value="$PROJECT_DIR$/ui_atoms" />
           </set>
         </option>
