+++ app/src/main/java/com/example/expensestracker/di/MediatorsBindings.kt
@@ -3,7 +3,9 @@ package com.example.expensestracker.di
 import com.example.add_transaction.navigation.AddTransactionScreenNavigatorImpl
 import com.example.core_api.mediators.AddTransactionScreenNavigator
 import com.example.core_api.mediators.GeneralScreenNavigator
+import com.example.core_api.mediators.TransactionsScreenNavigator
 import com.example.general.navigation.GeneralScreenNavigatorImpl
+import com.example.transactions.navigation.TransactionsScreenNavigatorImpl
 import dagger.Binds
 import dagger.Module
 import dagger.Reusable
@@ -22,4 +24,10 @@ interface MediatorsBindings {
     fun bindsAddTransactionMediator(
         addTransactionMediatorImpl: AddTransactionScreenNavigatorImpl
     ): AddTransactionScreenNavigator
+
+    @Binds
+    @Reusable
+    fun bindsTransactionsMediator(
+        transactionsMediatorImpl: TransactionsScreenNavigatorImpl
+    ): TransactionsScreenNavigator
 }
