+++ core_api/src/main/java/com/example/core_api/mediators/MediatorsProvider.kt
@@ -5,4 +5,6 @@ interface MediatorsProvider {
     fun provideGeneralMediator(): GeneralScreenNavigator
 
     fun provideAddTransactionMediator(): AddTransactionScreenNavigator
+
+    fun provideTransactionsMediator(): TransactionsScreenNavigator
 }
