+++ core_api/src/main/java/com/example/core_api/mediators/TransactionsScreenNavigator.kt
@@ -1,0 +1,9 @@
+package com.example.core_api.mediators
+
+import androidx.annotation.IdRes
+import androidx.fragment.app.FragmentManager
+
+interface TransactionsScreenNavigator {
+
+    fun startTransactionScreen(@IdRes containerId: Int, fragmentManager: FragmentManager)
+}
