+++ settings.gradle.kts
@@ -28,3 +28,4 @@ include(":main")
 include(":general")
 include(":ui_atoms")
 include(":add_transaction")
+include(":transactions")
