+++ transactions/src/main/java/com/example/transactions/models/TransactionCategory.kt
@@ -1,0 +1,6 @@
+package com.example.transactions.models
+
+data class TransactionCategory(
+    val title: String,
+    val amount: Int
+)
