+++ transactions/src/main/java/com/example/transactions/views/TransactionPeriodType.kt
@@ -1,0 +1,18 @@
+package com.example.transactions.views
+
+enum class TransactionPeriodType {
+    DAY,
+    WEEK,
+    MONTH,
+    YEAR,
+    PERIOD;
+
+    val dateFormat: String
+        get() = when (this) {
+            DAY -> "EEE, dd MMMM yyyy"
+            WEEK -> "dd MMM"
+            MONTH -> "MMMM yyyy"
+            YEAR -> "yyyy"
+            PERIOD -> "dd MMM yyyy"
+        }
+}
