+++ transactions/src/main/res/drawable/ic_diagram.xml
@@ -1,0 +1,22 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="60"
+    android:viewportHeight="60">
+
+    <path
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:pathData="M60,32c-0.474,6.588 -4.111,13.6 -7.5,18L34,32h26Z" />
+
+    <path
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:pathData="M48.432,53.732A30.047,30.047 0,0 1,0 29.92c0,-15.7 11.48,-28.536 28,-29.92v32Z" />
+
+    <path
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:pathData="M32,0a30.058,30.058 0,0 1,28 28L32,28L32,0Z" />
+
+</vector>
