+++ transactions/src/main/res/values/strings.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="transactions_period_day">День</string>
+    <string name="transactions_period_week">Неделя</string>
+    <string name="transactions_period_month">Месяц</string>
+    <string name="transactions_period_year">Год</string>
+    <string name="transactions_period_period">Период</string>
+    <string name="transactions_period_all_time">Все время</string>
+    <string name="transactions_period_hyphen">—</string>
+
+    <string name="transactions_total_amount_title">Всего:</string>
+
+    <string name="add_expense_title">Расход</string>
+</resources>
