+++ ui_atoms/src/main/res/drawable/arrow_dropdown_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_arrow_drop_up" android:state_checked="true" />
+    <item android:drawable="@drawable/ic_arrow_drop_down" />
+</selector>
