+++ ui_atoms/src/main/res/drawable/ic_arrow_left.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="12dp"
+    android:height="12dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+
+    <path
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:pathData="M14.11,20L15.5,18.594L7.239,9.987L8.12,9.069L8.115,9.074L15.455,1.427L14.086,0C12.057,2.113 6.393,8.014 4.5,9.987C5.906,11.453 4.535,10.024 14.11,20"
+        android:strokeWidth="1"
+        android:strokeColor="#00000000" />
+
+</vector>
