+++ ui_atoms/src/main/res/drawable/ic_arrow_right.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="12dp"
+    android:height="12dp"
+    android:viewportWidth="20"
+    android:viewportHeight="20">
+
+    <path
+        android:fillColor="#ffffff"
+        android:fillType="evenOdd"
+        android:pathData="M5.89,0L4.5,1.406L12.761,10.013L11.88,10.931L11.885,10.926L4.545,18.573L5.914,20C7.943,17.887 13.607,11.986 15.5,10.013C14.094,8.547 15.465,9.976 5.89,0"
+        android:strokeWidth="1"
+        android:strokeColor="#00000000" />
+
+</vector>
